Organize code: Move black and white printer to a "bw" package
