+++ src/main/java/com/diogonunes/jcdp/color/api/IColoredPrinter.java
@@ -1,6 +1,6 @@
 package com.diogonunes.jcdp.color.api;
 
-import com.diogonunes.jcdp.api.IPrinter;
+import com.diogonunes.jcdp.bw.api.IPrinter;
 import com.diogonunes.jcdp.color.api.Ansi.Attribute;
 import com.diogonunes.jcdp.color.api.Ansi.BColor;
 import com.diogonunes.jcdp.color.api.Ansi.FColor;
